Restructure UI Base and Filters folders and relocate CloseAllLogsAction to EventLog
