+++ src/EventLogExpert.Runtime/EventLog/CloseAllLogsAction.cs
@@ -1,6 +1,6 @@
 // // Copyright (c) Microsoft Corporation.
 // // Licensed under the MIT License.
 
-namespace EventLogExpert.Runtime.Common.Lifecycle;
+namespace EventLogExpert.Runtime.EventLog;
 
 public sealed record CloseAllLogsAction;
