+++ src/EventLogExpert.Runtime/LogTable/Reducers.cs
@@ -4,7 +4,7 @@
 using EventLogExpert.Eventing.Common.Channels;
 using EventLogExpert.Eventing.Common.EventLogs;
 using EventLogExpert.Eventing.Common.Events;
-using EventLogExpert.Runtime.Common.Lifecycle;
+using EventLogExpert.Runtime.EventLog;
 using Fluxor;
 using System.Collections.Immutable;
