+++ src/EventLogExpert.UI/FilterEditor/ComparisonOperatorSelect.razor
@@ -1,4 +1,4 @@
-@namespace EventLogExpert.UI.Filters
+@namespace EventLogExpert.UI.FilterEditor
 
 <ValueSelect AriaLabel="@(AriaLabelledBy is null ? AriaLabel ?? "Comparison" : null)"
     AriaLabelledBy="@AriaLabelledBy"
