+++ src/EventLogExpert.UI/FilterEditor/FilterRow.razor.cs
@@ -1,7 +1,7 @@
 // // Copyright (c) Microsoft Corporation.
 // // Licensed under the MIT License.
 
-using EventLogExpert.UI.Filters.Base;
+using EventLogExpert.UI.FilterEditor;
 using EventLogExpert.Filtering.Drafts;
 using EventLogExpert.Filtering.Persistence;
 using EventLogExpert.Filtering.Runtime;
@@ -12,7 +12,7 @@ using EventLogExpert.Runtime.FilterPane;
 using Fluxor;
 using Microsoft.AspNetCore.Components;
 
-namespace EventLogExpert.UI.Filters;
+namespace EventLogExpert.UI.FilterEditor;
 
 public sealed partial class FilterRow : FilterRowBase<SavedFilter?>
 {
