+++ src/EventLogExpert.UI/FilterEditor/FilterRow.razor.css
