+++ src/EventLogExpert.UI/FilterEditor/FilterRowChrome.razor
