+++ src/EventLogExpert.UI/FilterEditor/SubFilterRow.razor
