+++ src/EventLogExpert.UI/Inputs/TextInput.razor.cs
@@ -1,7 +1,7 @@
 // // Copyright (c) Microsoft Corporation.
 // // Licensed under the MIT License.
 
-using EventLogExpert.UI.Base;
+using EventLogExpert.UI.Inputs;
 using Microsoft.AspNetCore.Components;
 
 namespace EventLogExpert.UI.Inputs;
