+++ src/EventLogExpert.UI/Modal/ModalChrome.razor
