+++ src/EventLogExpert.UI/_Imports.razor
@@ -4,13 +4,11 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using EventLogExpert.UI.Alerts
 @using EventLogExpert.UI.Banner
-@using EventLogExpert.UI.Base
 @using EventLogExpert.UI.Database
 @using EventLogExpert.UI.DebugLog
 @using EventLogExpert.UI.FilterCache
 @using EventLogExpert.UI.FilterGroup
-@using EventLogExpert.UI.Filters
-@using EventLogExpert.UI.Filters.Base
+@using EventLogExpert.UI.FilterEditor
 @using EventLogExpert.UI.Inputs
 @using EventLogExpert.UI.Menu
 @using EventLogExpert.UI.Modal
