+++ src/EventLogExpert/Services/MauiMenuActionService.cs
@@ -10,9 +10,8 @@ using EventLogExpert.Eventing.Readers;
 using EventLogExpert.Platforms.Windows;
 using EventLogExpert.Runtime.Alerts;
 using EventLogExpert.Runtime.Common.Clipboard;
-using EventLogExpert.Runtime.Common.Lifecycle;
-using EventLogExpert.Runtime.Common.Versioning;
 using EventLogExpert.Runtime.EventLog;
+using EventLogExpert.Runtime.Common.Versioning;
 using EventLogExpert.Runtime.FilterPane;
 using EventLogExpert.Runtime.Menu;
 using EventLogExpert.Runtime.Modal;
