+++ src/EventLogExpert/_Imports.razor
@@ -8,14 +8,12 @@
 @using EventLogExpert.UI
 @using EventLogExpert.UI.Alerts
 @using EventLogExpert.UI.Banner
-@using EventLogExpert.UI.Base
 @using EventLogExpert.UI.Database
 @using EventLogExpert.UI.DebugLog
 @using EventLogExpert.UI.ErrorHandling
 @using EventLogExpert.UI.FilterCache
 @using EventLogExpert.UI.FilterGroup
-@using EventLogExpert.UI.Filters
-@using EventLogExpert.UI.Filters.Base
+@using EventLogExpert.UI.FilterEditor
 @using EventLogExpert.UI.Inputs
 @using EventLogExpert.Components.Layout
 @using EventLogExpert.UI.Menu
